+++ library/laboratory/src/main/java/io/mehow/laboratory/FeatureFactory.kt
@@ -13,7 +13,6 @@ public interface FeatureFactory {
   /**
    * Creates a new [FeatureFactory] that will return a combined set of this factory and the other factory.
    */
-  @JvmDefault
   public operator fun plus(factory: FeatureFactory): FeatureFactory = object : FeatureFactory {
     override fun create() = this@FeatureFactory.create() + factory.create()
   }